Bonus -- Part 3: Trade Parser Unit Tests
One of the advantages of this final version with multiple classes and interfaces is that it makes unit tests easier to create. You must create unit tests for the SimpleTradeParser class. Remember that the SimpleTradeParse has been refactored to use a trade validator and a trade mapper:

We will need to create dummy or mock versions of a trade validator and trade mapper for the parser to use. Here is one way to do this: sample mock classes

Try to create at least two unit tests:

1. A unit test that processes a number of strings which are all validated positively.
2. A unit test where the strings are validated negatively.