Asg 10 - Async Interfaces
Part 1: Async Updates

Change interfaces to Async
Use IAsyncEnumerable
From ChatGPT:

Directly Using async in the Interface (if allowed)

If the ITradeDataProvider interface could be modified to include asynchronous methods, you could make GetTradeData return Task<IEnumerable<string>>. This approach would avoid blocking and make full use of asynchronous operations.
csharppublic interface ITradeDataProvider
{
    Task<IEnumerable<string>> GetTradeDataAsync();
}
Then GetTradeDataAsync would call GetTradeAsync directly with await, without using Task.Run or blocking.

Details from Tom:

We are going to make the ITradeDataProvider interface asynchronous so it will now have

Task<IEnumerable<string>> GetTradeData();
instead of

IEnumerable<string> GetTradeData();
This is because some of the trade providers are synchronous and some are asynchronous. In these cases, it generally better to make the interface asynchronous and trick the synchronous methods to act like an async.

See Issues with an interface treating a synchronous action as async.

When we convert a synchronous method so that it will pretend to be asynchronous, we will use Task.FromResult( ) to return a Task wrapped around the returned object. See What is the use for Task.FromResult in C# for more.

Steps: (Coding trouble?  Add a description of goal and the trouble code)

1. Change the ITradeDataProvider interface so that GetTradeData() returns a Task--see above.
2. Update AsyncURLProvider  so that GetTradeData() returns a task and does not wait for the result.
3. Update RestfulTradeDataProvider so that GetTradeData() returns either
GetProductAsync();
Task.FromResult(emptyList);
4. Update StreamTradeDataProvider and URLTradeDataProvider so they use Task.FromResult() to return a Task.
5. Fix the error now in the TradeProcessor class by using await to call GetTradeData(). You may have to make a method async to do this.
6. Bonus -- Fix the errors in the unit tests by using await when calling the trade providers.